+++ tests/scheduler/test_pulse_library.py
@@ -9,6 +9,7 @@ import pytest
 
 import numpy as np
 from quantify_scheduler import Operation, Schedule
+from quantify_scheduler.backends import SerialCompiler
 from quantify_scheduler.json_utils import ScheduleJSONDecoder, ScheduleJSONEncoder
 from quantify_scheduler.operations.gate_library import X90, X
 from quantify_scheduler.operations.pulse_library import (
@@ -23,11 +24,7 @@ from quantify_scheduler.operations.pulse_library import (
     create_dc_compensation_pulse,
     decompose_long_square_pulse,
 )
-from quantify_scheduler.backends import SerialCompiler
 from quantify_scheduler.resources import BasebandClockResource, ClockResource
-from quantify_scheduler.device_under_test.transmon_element import (
-    BasicTransmonElement,
-)
 
 
 # --------- Test classes and member methods ---------
@@ -378,19 +375,21 @@ def test_dccompensation_pulse_both_params() -> None:
 
 # --------- Test pulse compilation ---------
 def test_dragpulse_motzoi(mock_setup_basic_transmon):
-
     mock_setup_basic_transmon["q0"].rxy.amp180(0.2)
     mock_setup_basic_transmon["q0"].rxy.motzoi(0.02)
 
     sched = Schedule("Test DRAG Pulse")
     sched.add(X("q0"))
+
     compiler = SerialCompiler(name="compiler")
     compiled_sched = compiler.compile(
         sched, mock_setup_basic_transmon["quantum_device"].generate_compilation_config()
     )
-    for op, op_info in compiled_sched.operations.items():
-        D_amp = op_info.data["pulse_info"][0].get("D_amp")
 
-    assert (
-        D_amp == mock_setup_basic_transmon["q0"].rxy.motzoi()
-    ), "The amplification of the derivative DRAG pulse is not equal to the motzoi parameter"
+    D_amp = (
+        list(compiled_sched.operations.values())[0].data["pulse_info"][0].get("D_amp")
+    )
+    assert D_amp == mock_setup_basic_transmon["q0"].rxy.motzoi(), (
+        "The amplification of the derivative DRAG pulse is not equal to the motzoi "
+        "parameter"
+    )